全体のコンポーネント化
